Implementing Collin's feedback
