+++ .gitignore
@@ -158,13 +158,4 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 .idea/
-examples/.DS_Store
-examples/file_extras/.DS_Store
-examples/file_extras/tools/.DS_Store
-examples/file_extras/tools/metadata_tool/.DS_Store
-examples/file_extras/tools/open_file_tool/.DS_Store
-griptape/file_extras/tools/.DS_Store
-tests/unit/.DS_Store
-tests/unit/tools/.DS_Store
-tests/unit/tools/metadata_tool/.DS_Store
-tests/unit/tools/open_file_tool/.DS_Store
+.DS_Store
