+++ docs/file_extras/tools/open_file_tool/README.md
@@ -8,25 +8,22 @@ For example, ask an agent to open an image from /path/to/image and it will displ
 from dotenv import load_dotenv
 
 from griptape.file_extras.tools.open_file_tool import OpenFileTool
+from griptape.file_extras.utils.env_utils import get_env_var
 from griptape.structures import Agent
 from griptape.drivers import OpenAiChatPromptDriver
 
-import os
-
 load_dotenv()
 
-key = "OPENAI_API_KEY"
-if not (OPENAI_API_KEY := os.getenv(key)):
-    raise ValueError(f"{key} missing")
+api_key = get_env_var( "OPENAI_API_KEY" )
 
 TEST_IMAGE = "examples/file_extras/tools/open_file_tool/media/capybara_cloud.jpeg"
 
 agent = Agent(
-    prompt_driver=OpenAiChatPromptDriver(model="gpt-4o-mini", api_key=OPENAI_API_KEY),
+    prompt_driver=OpenAiChatPromptDriver(model="gpt-4o-mini", api_key=api_key),
     tools=[OpenFileTool()],
 )
 
-agent.run( f"Can you show me the image at {TEST_IMAGE}?" )
+agent.run(f"Can you show me the image at {TEST_IMAGE}?")
 ```
 
 will output:
