+++ examples/file_extras/tools/open_file_tool/example_agent.py
@@ -1,21 +1,18 @@
 from dotenv import load_dotenv
 
 from griptape.file_extras.tools.open_file_tool import OpenFileTool
+from griptape.file_extras.utils.env_utils import get_env_var
 from griptape.structures import Agent
 from griptape.drivers import OpenAiChatPromptDriver
 
-import os
-
 load_dotenv()
 
-key = "OPENAI_API_KEY"
-if not (OPENAI_API_KEY := os.getenv(key)):
-    raise ValueError(f"{key} missing")
+api_key = get_env_var("OPENAI_API_KEY")
 
 TEST_IMAGE = "examples/file_extras/tools/open_file_tool/media/capybara_cloud.jpeg"
 
 agent = Agent(
-    prompt_driver=OpenAiChatPromptDriver(model="gpt-4o-mini", api_key=OPENAI_API_KEY),
+    prompt_driver=OpenAiChatPromptDriver(model="gpt-4o-mini", api_key=api_key),
     tools=[OpenFileTool()],
 )
