+++ pyproject.toml
@@ -12,7 +12,6 @@ packages = [
 python = "^3.9"
 griptape = "^1.0.0"
 python-dotenv = "^1.0.1"
-pyright = "^1.1.390"
 
 [tool.poetry.group.test]
 optional = true
